+++ src/app/Fake.SQL/SQLServer.fs
@@ -118,7 +118,16 @@ let runScript serverInfo sqlFile =
       |> (getDatabase serverInfo).ExecuteNonQuery
     
 /// Closes the connection to the server
-let Disconnect serverInfo = serverInfo.Server.ConnectionContext.Disconnect()
+let Disconnect serverInfo = 
+    logfn "Disconnecting from server %s." (getServerName serverInfo)
+    try
+        if serverInfo.Server = null then
+            failwith "Server is not configured"
+        if serverInfo.Server.ConnectionContext = null then
+            failwith "Server.ConnectionContext is not configured"
+        serverInfo.Server.ConnectionContext.Disconnect()
+    with
+    | e -> traceImportant <| sprintf "Disconnecting failed with %s" e.Message
 
 /// Replaces the database files
 let internal replaceDatabaseFiles connectionString attachOptions copyF =